demande prolongement

+++ src/main/java/com/example/demo/repository/AdherentRepository.java
@@ -55,5 +55,5 @@ public interface AdherentRepository extends JpaRepository<Adherent, Long> {
 
        // getQuotaProlongement
        @Query("SELECT a.typeAdherent.quotaProlongement FROM Adherent a WHERE a.id = :idAdherent")
-       int getQuotaProlongement(@Param("idAdherent") int idAdherent);
+       int getQuotaProlongement(@Param("idAdherent") Long idAdherent);
 }
